// old/ProcureHub.WebApp/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        QueryDepartmentsResponse: {
            /** Format: uuid */
            id: string;
            name: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}